+++ xenum/lib/inc/xenum5/impl/CstringDeclValue.hpp
@@ -17,7 +17,9 @@
  */
 // FIXME: ext/cxp
 #define _XENUM5_CSTRING_DECLV(PNAME, DEPTH, PDEF, CTXT, Z)					\
-	BOOST_PP_CAT(BOOST_PP_CAT(_XENUM5_CSTRING_, _XENUM5_PDEF_PLACEMENT_STR(PDEF)), _DECLV)	\
+	/* INC() to ensure that the index type always gets defined */				\
+	_XENUM5_PROP_DECLV_TYPES(PNAME, BOOST_PP_INC(DEPTH))					\
+	BOOST_PP_CAT(_XENUM5_CSTRING_DECLV_, _XENUM5_PDEF_PROP_DATA(PDEF))			\
 		(PNAME, DEPTH, PDEF, CTXT, Z)							\
 
 
@@ -25,36 +27,17 @@
 /**
  * Declare the functions related to a single custom property, implemented in header.
  */
-#define _XENUM5_CSTRING_HDR_DECLV(PNAME, DEPTH, PDEF, CTXT, Z)					\
-	/* INC() to ensure that the index type always gets defined */				\
-	_XENUM5_PROP_DECLV_TYPES(PNAME, BOOST_PP_INC(DEPTH))					\
-	_XENUM5_CSTRING_HDR_DECLV_FUNCS(PNAME, DEPTH, PDEF, CTXT, Z)				\
+#define _XENUM5_CSTRING_DECLV_HDR(PNAME, DEPTH, PDEF, CTXT, Z)					\
+	/* INC() because IndexNodes also has indexnodes for the leaf string values */		\
+	_XENUM5_PROP_HDR_DECLV_GET_SIZE(BOOST_PP_INC(DEPTH), PDEF, Z)				\
+	_XENUM5_PROP_HDR_DECLV_GET_VALUE(PNAME, DEPTH, PDEF, Z)					\
 
 
 // ======================================= MAIN (SRC) ===========================================
 /**
  * Declare the data related to a single custom property, implemented in source.
  */
-#define _XENUM5_CSTRING_SRC_DECLV(PNAME, DEPTH, PDEF, CTXT, Z)					\
-	_XENUM5_PROP_DECLV_TYPES(PNAME, BOOST_PP_INC(DEPTH))					\
-	_XENUM5_CSTRING_SRC_DECLV_FUNCS(PNAME, DEPTH, PDEF, CTXT, Z)				\
-
-
-// ===================================== FUNCTIONS (HDR) ========================================
-/**
- * Worker for _XENUM5_PROP_DECLV_PLAIN().
- * Declare the functions related to a single custom property, implemented in header.
- */
-#define _XENUM5_CSTRING_HDR_DECLV_FUNCS(PNAME, DEPTH, PDEF, CTXT, Z)				\
-	_XENUM5_PROP_HDR_DECLV_GET_SIZE(BOOST_PP_INC(DEPTH), PDEF, Z)				\
-	_XENUM5_PROP_HDR_DECLV_GET_VALUE(PNAME, DEPTH, PDEF, Z)					\
-
-
-// ===================================== FUNCTIONS (SRC) ========================================
-/**
- * Declare the functions related to a single custom property, in value class.
- */
-#define _XENUM5_CSTRING_SRC_DECLV_FUNCS(PNAME, DEPTH, PDEF, CTXT, Z)				\
+#define _XENUM5_CSTRING_DECLV_SRC(PNAME, DEPTH, PDEF, CTXT, Z)					\
 	/* INC() because IndexNodes also has indexnodes for the leaf string values */		\
 	_XENUM5_PROP_SRC_DECLV_GET_SIZE(BOOST_PP_INC(DEPTH), PDEF, Z)				\
 	_XENUM5_PROP_SRC_DECLV_GET_VALUE(PNAME, DEPTH, PDEF, Z)					\